-refactor WebDriverInitializerImpl; -move extension creation logic to ChromeProxyPlugin; -add browser version and chrome path properties
