+++ src/main/java/executor/service/config/properties/PropertiesConstants.java
@@ -17,11 +17,11 @@ public final class PropertiesConstants {
     public static final String KEEP_ALIVE_TIME = "keep.alive.time";
     public static final String COMMONS_CONFIGURATION_PROPERTIES = "web-driver.properties";
     public static final String PROXY_VALIDATOR_TARGET_URL = "target.url";
-    public static final Integer PROXY_VALIDATOR_CONNECTION_TIMEOUT = Integer.getInteger("connection.timeout");
-
-
+    public static final Integer PROXY_VALIDATOR_CONNECTION_TIMEOUT = Integer.parseInt("connection.timeout");
     public static final String WEB_DRIVER = "web-driver.properties";
     public static final String WEB_DRIVER_EXECUTABLE = "webDriver-executable";
+    public static final String CHROME_EXECUTABLE = "chrome-executable";
+    public static final String CHROME_VERSION = "chrome-version";
     public static final String USER_AGENT = "user-agent";
     public static final String PAGE_LOAD_TIMEOUT = "page-load-timeout";
     public static final String IMPLICITLY_WAIT = "implicitly-wait";
